requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
choicelib==0.1.4
idna==3.2
multidict==5.1.0
Pillow==8.3.2
pydantic==1.8.2
python-dotenv==0.19.0
typing-extensions==3.10.0.2
vbml==1.1
vkbottle==3.0.2
vkbottle-types==5.131
watchgod==0.7
yarl==1.6.3
